+++ liquidhaskell.cabal
@@ -161,13 +161,17 @@ Library
                     Language.Haskell.Liquid.GHC.Interface,
                     Language.Haskell.Liquid.Types.SpanStack,
                     Language.Haskell.Liquid.Types.RefType,
-                    Language.Haskell.Liquid.Transforms.RefSplit,
-                    Language.Haskell.Liquid.UX.Errors,
+                    Language.Haskell.Liquid.Types.Errors,
                     Language.Haskell.Liquid.Types.Specifications,
                     Language.Haskell.Liquid.Types.PredType,
                     Language.Haskell.Liquid.UX.ACSS,
                     Language.Haskell.Liquid.UX.DiffCheck,
                     Language.Haskell.Liquid.Transforms.ANF,
+                    Language.Haskell.Liquid.Transforms.RefSplit,
+                    Language.Haskell.Liquid.Transforms.CoreToLogic,
+                    Language.Haskell.Liquid.Transforms.Rec,
+                    Language.Haskell.Liquid.Transforms.Simplify,
+                    Language.Haskell.Liquid.UX.Errors,
                     Language.Haskell.Liquid.UX.Annotate,
                     Language.Haskell.Liquid.UX.CTags,
                     Language.Haskell.Liquid.UX.Config,
@@ -176,15 +180,12 @@ Library
                     Language.Haskell.Liquid.GHC.Misc,
                     Language.Haskell.Liquid.GHC.Play,
                     Language.Haskell.Liquid.Misc,
-                    Language.Haskell.Liquid.Transforms.CoreToLogic,
                     Language.Haskell.Liquid.Types.Variance,
                     Language.Haskell.Liquid.Types.Bounds,
                     Language.Haskell.Liquid.Types.Dictionaries,
                     Language.Haskell.Liquid.Constraint.Qualifier,
-                    Language.Haskell.Liquid.Transforms.Rec,
                     Language.Haskell.Liquid.UX.Tidy,
                     Language.Haskell.Liquid.Types,
-                    Language.Haskell.Liquid.Transforms.Simplify,
                     Language.Haskell.Liquid.Types.Literals,
                     Language.Haskell.Liquid.Types.Strata,
                     Language.Haskell.Liquid.Constraint.Fresh,
